CDFDatasets

CDFDatasets.jl is a Julia package for reading CDF (Common Data Format) files, commonly used in space physics and other scientific domains. It provides a Julia interface to CDF files using the CommonDataModel.jl interface.

Installation

using Pkg
Pkg.add("CDFDatasets")

Quickstart

Here's a quick example using OMNI solar wind data:

using CDFDatasets

# Open a CDF dataset
omni_file = joinpath(pkgdir(CDFDatasets), "data/omni_coho1hr_merged_mag_plasma_20240901_v01.cdf")
ds = CDFDataset(omni_file)

Dataset: 
Group: /

Variables
  Epoch   (720)
    Datatype:    Dates.DateTime
    Dimensions:  
    Attributes:
     FIELDNAM             = Epoch Time
     CATDESC              = Epoch Time
     VALIDMIN             = Any[<py 1963-01-01T00:00:00.000000000>]
     VALIDMAX             = Any[<py 2020-12-31T23:59:59.000000000>]
     SCALEMIN             = Any[<py 1963-01-01T00:00:00.000000000>]
     SCALEMAX             = Any[<py 2020-12-31T23:59:59.000000000>]
     UNITS                = DD-MMM-YYYY_hr:mm
     FILLVAL              = Any[-1.0e31]
     VAR_TYPE             = support_data

  heliographicLatitude   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = Heliographic Latitude
     CATDESC              = HelioGraphic Inertial (HGI) latitude
     VALIDMIN             = Any[-90.0]
     VALIDMAX             = Any[90.0]
     UNITS                = deg
     FORMAT               = f7.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = Heliographic Latitude

  heliographicLongitude   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = Heliographic Longitude
     CATDESC              = HGI longitude
     VALIDMIN             = Any[0.0]
     VALIDMAX             = Any[360.0]
     UNITS                = deg
     FORMAT               = f7.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = Heliographic Longitude

  BR   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = BR (RTN)
     CATDESC              = BR in RTN (Radial-Tangential-Normal) coordinate system
     VALIDMIN             = Any[-1000.0]
     VALIDMAX             = Any[1000.0]
     UNITS                = nT
     FORMAT               = f6.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = BR (RTN)

  BT   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = BT (RTN)
     CATDESC              = BT in RTN coordinate system
     VALIDMIN             = Any[-1000.0]
     VALIDMAX             = Any[1000.0]
     UNITS                = nT
     FORMAT               = f6.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = BT (RTN)

  BN   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = BN (RTN)
     CATDESC              = BN in RTN coordinate system
     VALIDMIN             = Any[-1000.0]
     VALIDMAX             = Any[1000.0]
     UNITS                = nT
     FORMAT               = f6.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = BN (RTN)

  ABS_B   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = Field Magnitude Avg.
     CATDESC              = Field Magnitude Average |B  1/N SUM |B|
     VALIDMIN             = Any[0.0]
     VALIDMAX             = Any[100.0]
     UNITS                = nT
     FORMAT               = F6.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = Field Magnitude Avg.

  V   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = Bulk Flow Speed
     CATDESC              = Bulk Flow Speed
     VALIDMIN             = Any[0.0]
     VALIDMAX             = Any[1200.0]
     UNITS                = km/s
     FORMAT               = F5.0
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = Bulk Flow Speed

  elevAngle   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = elevation Angle 
     CATDESC              = Proton flow elevation angle / latitude (RTN)
     VALIDMIN             = Any[-90.0]
     VALIDMAX             = Any[90.0]
     UNITS                = Deg
     FORMAT               = F6.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = elevation Angle

  azimuthAngle   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = azimuth Angle 
     CATDESC              = Proton flow azimuth angle / longitude (RTN)
     VALIDMIN             = Any[-180.0]
     VALIDMAX             = Any[180.0]
     UNITS                = Deg
     FORMAT               = F6.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = azimuth Angle

  N   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = Ion Density
     CATDESC              = Ion Density
     VALIDMIN             = Any[0.0]
     VALIDMAX             = Any[150.0]
     UNITS                = N/cm3
     FORMAT               = F6.1
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = Ion density

  T   (720)
    Datatype:    Float32
    Dimensions:  Epoch
    Attributes:
     FIELDNAM             = Temperature
     CATDESC              = Temperature
     VALIDMIN             = Any[0.0]
     VALIDMAX             = Any[1.0e7]
     UNITS                = Deg K
     FORMAT               = F9.0
     FILLVAL              = Any[-9.999999848243207e30]
     VAR_TYPE             = data
     DISPLAY_TYPE         = time_series
     DEPEND_0             = Epoch
     LABLAXIS             = Temperature

Global attributes
  Project              = SPDF
  Discipline           = Space Physics>Interplanetary Studies
  Source_name          = OMNI (1AU IP Data)>Merged 1 Hour Interplantary OMNI data in RTN system
  Data_type            = COHO1HR>Definitive Hourly data from cohoweb
  Descriptor           = merged magnetic field and plasma data from cohoweb
  Data_version         = 1
  TITLE                = Near-Earth Heliosphere Data (OMNI)
  TEXT                 = Hourly averaged definitive multispacecraft interplanetary parameters data
  MODS                 = created July 2007;
  PI_name              = J.H. King, N. Papatashvilli
  PI_affiliation       = AdnetSystems, NASA GSFC
  Generation_date      = Ongoing
  Acknowledgement      = NSSDC
  ADID_ref             = NSSD0110
  Rules_of_use         = Public
  Instrument_type      = Plasma and Solar Wind
  Generated_by         = King/Papatashvilli
  Time_resolution      = 1 hour
  Logical_file_id      = omni_coho1hr_merged_mag_plasma_00000000_v01
  Logical_source       = omni_coho1hr_merged_mag_plasma
  Logical_source_description = OMNI Combined merged hourly magnetic field, plasma and ephermis data
  LINK_TEXT            = COHO dataset
  LINK_TITLE           = Documentation
  HTTP_LINK            = https://omniweb.gsfc.nasa.gov/coho/html/cw_data.html
  alt_logical_source   = Combined_OMNI_1AU-MagneticField-Plasma-Particles_mrg1hr_1hour_cdf
  Mission_group        = OMNI (Combined 1AU IP Data; Magnetic and Solar Indices)
  spase_DatasetResourceID = spase://NASA/NumericalData/OMNI/COHO/MergedMagPlasma/PT1H
  DOI                  = https://doi.org/10.48322/6ffx-3441

Explore the dataset

julia> println("Variables: ", keys(ds))Variables: ["Epoch", "heliographicLatitude", "heliographicLongitude", "BR", "BT", "BN", "ABS_B", "V", "elevAngle", "azimuthAngle", "N", "T"]
julia> println("Attributes: ", keys(ds.attrib))Attributes: ["Project", "Discipline", "Source_name", "Data_type", "Descriptor", "Data_version", "TITLE", "TEXT", "MODS", "PI_name", "PI_affiliation", "Generation_date", "Acknowledgement", "ADID_ref", "Rules_of_use", "Instrument_type", "Generated_by", "Time_resolution", "Logical_file_id", "Logical_source", "Logical_source_description", "LINK_TEXT", "LINK_TITLE", "HTTP_LINK", "alt_logical_source", "Mission_group", "spase_DatasetResourceID", "DOI"]
julia> ds.attrib["Descriptor"]"merged magnetic field and plasma data from cohoweb"

Access variables

julia> ds["Epoch"]Epoch (720)
  Datatype:    Dates.DateTime
  Dimensions:
  Attributes:
   FIELDNAM             = Epoch Time
   CATDESC              = Epoch Time
   VALIDMIN             = Any[<py 1963-01-01T00:00:00.000000000>]
   VALIDMAX             = Any[<py 2020-12-31T23:59:59.000000000>]
   SCALEMIN             = Any[<py 1963-01-01T00:00:00.000000000>]
   SCALEMAX             = Any[<py 2020-12-31T23:59:59.000000000>]
   UNITS                = DD-MMM-YYYY_hr:mm
   FILLVAL              = Any[-1.0e31]
   VAR_TYPE             = support_data
julia> ds["Epoch"][[1,end]]2-element Vector{Dates.DateTime}:
 2024-09-01T00:00:00
 2024-09-30T23:00:00
julia> ds["BR"]BR (720)
  Datatype:    Float32
  Dimensions:  Epoch
  Attributes:
   FIELDNAM             = BR (RTN)
   CATDESC              = BR in RTN (Radial-Tangential-Normal) coordinate system
   VALIDMIN             = Any[-1000.0]
   VALIDMAX             = Any[1000.0]
   UNITS                = nT
   FORMAT               = f6.1
   FILLVAL              = Any[-9.999999848243207e30]
   VAR_TYPE             = data
   DISPLAY_TYPE         = time_series
   DEPEND_0             = Epoch
   LABLAXIS             = BR (RTN)

# Calculate magnetic field magnitude
br = ds["BR"]
bt = ds["BT"]
bn = ds["BN"]
b_mag = sqrt.(br.^2 + bt.^2 + bn.^2) |> collect

720-element Vector{Float32}:
 11.184363
 10.812492
 10.683165
 10.7415085
 10.092076
  9.167878
  9.398936
  9.275236
  9.04931
  7.783958
  ⋮
  6.5015383
  6.497692
  5.9514704
  5.197115
  3.6496575
  4.532108
  6.603787
  5.5317264
  4.8394213

API Reference

CDFDatasets.CDFDataset

CDFDatasets.CDFDataset — Method

CDFDataset(file; lazy_load = true)

Load the CDF dataset at the file path. The dataset supports the API of the JuliaGeo/CommonDataModel.jl.

lazy_load controls whether variable values are loaded immediately or only when accessed by the user. If True, variables' values are loaded on demand. If False, all variable values are loaded during parsing.